# packages/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "KAIRO API",
    "description": "Languages, sandboxed executions, normalized diagnostics and Saarthi, the AI assistant.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/executions": {
      "post": {
        "tags": [
          "executions"
        ],
        "summary": "Create Execution",
        "description": "Queue a compile-and-run of the given source. Returns immediately.",
        "operationId": "create_execution_api_v1_executions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExecutionCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExecutionOut"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "413": {
            "description": "Request Entity Too Large",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/executions/{execution_id}": {
      "get": {
        "tags": [
          "executions"
        ],
        "summary": "Get Execution",
        "operationId": "get_execution_api_v1_executions__execution_id__get",
        "parameters": [
          {
            "name": "execution_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Execution Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExecutionOut"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/assistant/explain": {
      "post": {
        "tags": [
          "assistant"
        ],
        "summary": "Explain",
        "description": "Explain a diagnostic of a finished run, grounded in that run's evidence.",
        "operationId": "explain_api_v1_assistant_explain_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExplainIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExplanationOut"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "502": {
            "description": "Bad Gateway",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/assistant/fix": {
      "post": {
        "tags": [
          "assistant"
        ],
        "summary": "Fix",
        "description": "Propose the smallest patch for a diagnostic. The student must approve it.",
        "operationId": "fix_api_v1_assistant_fix_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FixIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FixOut"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "502": {
            "description": "Bad Gateway",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/assistant/verify": {
      "post": {
        "tags": [
          "assistant"
        ],
        "summary": "Verify",
        "description": "Decide from a real run of the patched code whether a fix worked.",
        "operationId": "verify_api_v1_assistant_verify_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerifyOut"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "502": {
            "description": "Bad Gateway",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/assistant/ask": {
      "post": {
        "tags": [
          "assistant"
        ],
        "summary": "Ask",
        "description": "Answer a question about programming, the editor's code or a run.",
        "operationId": "ask_api_v1_assistant_ask_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AskIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AskOut"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "502": {
            "description": "Bad Gateway",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/assistant/complete": {
      "post": {
        "tags": [
          "assistant"
        ],
        "summary": "Complete",
        "description": "Suggest what comes next at the cursor (the editor shows it as ghost text; Tab accepts).",
        "operationId": "complete_api_v1_assistant_complete_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompleteIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompleteOut"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "502": {
            "description": "Bad Gateway",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ErrorOut": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail"
          },
          "retryAfterS": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Retryafters"
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "ErrorOut"
      },
      "ExecutionCreate": {
        "properties": {
          "languageId": {
            "type": "string",
            "maxLength": 32,
            "minLength": 1,
            "title": "Languageid",
            "examples": [
              "c"
            ]
          },
          "source": {
            "type": "string",
            "title": "Source",
            "description": "Exact editor contents; its SHA-256 identifies this snapshot."
          },
          "stdin": {
            "type": "string",
            "title": "Stdin",
            "description": "Text given to the program on standard input.",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "languageId",
          "source"
        ],
        "title": "ExecutionCreate"
      },
      "ExecutionOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "languageId": {
            "type": "string",
            "title": "Languageid"
          },
          "state": {
            "$ref": "#/components/schemas/ExecutionState"
          },
          "terminal": {
            "type": "boolean",
            "title": "Terminal"
          },
          "summary": {
            "type": "string",
            "title": "Summary",
            "description": "One-line, plain-language status for the UI."
          },
          "sourceHash": {
            "type": "string",
            "title": "Sourcehash",
            "description": "sha256:<hex> of the exact source that was (or will be) compiled."
          },
          "sourceBytes": {
            "type": "integer",
            "title": "Sourcebytes"
          },
          "queuePosition": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Queueposition",
            "description": "1-based position while QUEUED."
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "title": "Createdat"
          },
          "startedAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Startedat"
          },
          "finishedAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finishedat"
          },
          "toolchain": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ToolchainOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "steps": {
            "items": {
              "$ref": "#/components/schemas/ExecutionStepOut"
            },
            "type": "array",
            "title": "Steps"
          },
          "diagnostics": {
            "items": {
              "$ref": "#/components/schemas/Diagnostic"
            },
            "type": "array",
            "title": "Diagnostics"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "id",
          "languageId",
          "state",
          "terminal",
          "summary",
          "sourceHash",
          "sourceBytes",
          "createdAt",
          "steps",
          "diagnostics"
        ],
        "title": "ExecutionOut"
      },
...
      "ExplainIn": {
        "properties": {
          "executionId": {
            "type": "string",
            "maxLength": 64,
            "title": "Executionid"
          },
          "sourceHash": {
            "type": "string",
            "pattern": "^sha256:[0-9a-f]{64}$",
            "title": "Sourcehash",
            "description": "Hash of the code in the editor; must match the run (no stale answers)."
          },
          "diagnosticId": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 128
              },
              {
                "type": "null"
              }
            ],
            "title": "Diagnosticid",
            "description": "Diagnostic to explain; default: the first error."
          },
          "question": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Question"
          }
        },
        "type": "object",
        "required": [
          "executionId",
          "sourceHash"
        ],
        "title": "ExplainIn"
      },
      "ExplanationOut": {
        "properties": {
          "executionId": {
            "type": "string",
            "title": "Executionid"
          },
          "sourceHash": {
            "type": "string",
            "title": "Sourcehash",
            "description": "The explanation applies to exactly this source."
          },
          "focusDiagnosticId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Focusdiagnosticid"
          },
          "diagnosticIds": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Diagnosticids",
            "description": "The diagnostics Saarthi was shown (its evidence)."
          },
          "problem": {
            "type": "string",
            "title": "Problem"
          },
          "whatHappened": {
            "type": "string",
            "title": "Whathappened"
          },
          "why": {
            "type": "string",
            "title": "Why"
          },
          "location": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CodeLocationOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "suggestedFix": {
            "type": "string",
            "title": "Suggestedfix"
          },
          "relatedConcepts": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Relatedconcepts"
          },
          "confidence": {
            "type": "string",
            "enum": [
              "low",
              "medium",
              "high"
            ],
            "title": "Confidence"
          },
          "verified": {
            "type": "boolean",
            "const": false,
            "title": "Verified",
            "description": "Explanations are advice; only a run verifies code.",
            "default": false
          },
          "provider": {
            "type": "string",
            "title": "Provider"
          },
          "model": {
            "type": "string",
            "title": "Model"
          },
          "cached": {
            "type": "boolean",
            "title": "Cached",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "executionId",
          "sourceHash",
          "focusDiagnosticId",
          "diagnosticIds",
          "problem",
          "whatHappened",
          "why",
          "location",
          "suggestedFix",
          "relatedConcepts",
          "confidence",
          "provider",
          "model"
        ],
        "title": "ExplanationOut"
      },
      "FixIn": {
        "properties": {
          "executionId": {
            "type": "string",
            "maxLength": 64,
            "title": "Executionid"
          },
          "sourceHash": {
            "type": "string",
            "pattern": "^sha256:[0-9a-f]{64}$",
            "title": "Sourcehash"
          },
          "diagnosticId": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 128
              },
              {
                "type": "null"
              }
            ],
            "title": "Diagnosticid"
          }
        },
        "type": "object",
        "required": [
          "executionId",
          "sourceHash"
        ],
        "title": "FixIn"
      },
      "FixOut": {
        "properties": {
          "fixId": {
            "type": "string",
            "title": "Fixid",
            "description": "Pass to /assistant/verify after running the patched code."
          },
          "executionId": {
            "type": "string",
            "title": "Executionid"
          },
          "baseSourceHash": {
            "type": "string",
            "title": "Basesourcehash",
            "description": "The patch applies to exactly this source."
          },
          "patchedSource": {
            "type": "string",
            "title": "Patchedsource"
          },
          "patchedSourceHash": {
            "type": "string",
            "title": "Patchedsourcehash"
          },
          "summary": {
            "type": "string",
            "title": "Summary"
          },
          "edits": {
            "items": {
              "$ref": "#/components/schemas/PatchEditOut"
            },
            "type": "array",
            "title": "Edits"
          },
          "targetCodes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Targetcodes",
            "description": "Diagnostic codes the fix is meant to remove."
          },
          "confidence": {
            "type": "string",
            "enum": [
              "low",
              "medium",
              "high"
            ],
            "title": "Confidence"
          },
          "verified": {
            "type": "boolean",
            "const": false,
            "title": "Verified",
            "description": "Not verified until the patched code is run.",
            "default": false
          },
          "provider": {
            "type": "string",
            "title": "Provider"
          },
          "model": {
            "type": "string",
            "title": "Model"
          },
          "cached": {
            "type": "boolean",
            "title": "Cached",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "fixId",
          "executionId",
          "baseSourceHash",
          "patchedSource",
          "patchedSourceHash",
          "summary",
          "edits",
          "targetCodes",
          "confidence",
          "provider",
          "model"
        ],
        "title": "FixOut"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
    }
  }
}